More housekeeping and some network analysis
After cleaning up the ORNLDAAC and E DSTORE IN directories I reran the
content distribution xsl. Will finish building dialect specific
quickies now that the records are in the proper directories.

Early attempts with Gephi to describe the DataOne network. Nodes in the
working resolution are dialects and member nodes. Edges are DataOne to
dialect and dialect to membernode, both undirected. This may be better
described directed with the target and sources reversed. Need to
incorporate paths as additional nodes and concepts being edges between
paths in dialects.