Content.QuickE_production
The first step towards this goal was to identify all of the versions of
the metadata dialects as each dialect was determined to be too low a
resolution when dealing with the early non canonical drafts. The
content distribution was too wide and multiple paths referred to the
same concepts. This ambiguity of the content of records written in the
dialect dissipates when refined to the version specific instantiation
of the dialect.

This commit is once the metadatacontent.txt files have been created for
each member node record sample set. These files are then concatenated
to be prepared for integration into a QuickE as the data worksheet.