Continued Network Analysis
These files have to do with describing DataOne as a network. Member
nodes are nodes and dialects that they share in common are edges
between the membernodes. This is a description of the ability of a
membernode to use metadata to communicate with other membernodes.
Additional files describe member nodes as sharing through a dialect to
DataOne.