HandHewnLvls
Applied the first three LTER recLvls to CSDGM QuickE. Most conceptual
paths had an easy real path to use. The only real problems are when a
dialect uses more than one way to contain a concept. I believe this
skews results. Should a single path be chosen out of the three? Example
is Contributor Name.

The other problem is specific to the Temporal Extent concept. With
single date and single time and begin and end date and begin and end
time many records that actually contain a temporal extent are not
counted.