Analysis_DialectRec
Finished building pipeline. Noticed that an xsl had issues. Rebuilt
son. Moved generateSS.py to run locally rather than performing that
part of the process in another environment.

I also created a “clean” xml directory. This removes funky namespaces
and recreates the directory structure using the dialect codes from
AllCrosswalks.

Future work to make the LTER/FGDC analyses of the EML and FGDC
stronger, the BDP and Mercury need to be integrated into concept
definitions for the concepts that occur in the above mentioned
recommendations.