FinalizeAnalyses
New sheets include: Concept Lists, Personal observations. Concept lists
detail the complete, missing, incomplete, and unused concepts as well
as the signature group specific missing elements. While these are
included in previous deliverables, they were not included in the
development until the presentation layer. Personal observations include
sheets that detail things about the data that I found interesting that
might not be immediately apparent.

I did not figure the signature group specific incomplete concepts
because there were too many signatures for it to be anything but data
supporting the assumption that while concepts in the LTER rec are
important to the organization that created it, they may not be as
important to other organizations conducting similar science. One of the
worksheets refines the signature groups by only including records from
LTER, LTER_EUROPE, and TERN and including the membernodes as columns
detailing which member node contains which signature groups.