CompleteConcepts_memeberNodes
There was a need identified for a new type of graph. As the analysis
has focused on dialect and recommendation, there were no quick answers
to this.

In an attempt to address the issue I developed a new worksheet. This
worksheet sums up the exists columns for each member node using a
specific dialect. These cells are also counted using the same
conditionals so that any cell at zero is counted as 1. The member node
columns are then counted if the number is bigger or equal to one. These
results are separated by levels and then graphed to fulfill the need
for visualizations.

This worksheet, with small variations can be applied to any other
analysis.