tablesTalk
filled in tables talked about earlier in the day, used the
statistic.csv the python sampler generates to arrive at the numbers.
Ended up including the sample entire, as there are only a couple of
nodes that don’t use one of the two. I also remember that we just used
csdgm for BDP and Mercury